+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Support static method calls with blocks in es2015+
 * Auto-bind instance methods referenced as properties within a class
-* handle begin as an expression
+* handle begin, if, and case as expressions
 
 # 3.6.1 / 2020-12-31